requirements.txt
# StellarMind - EarthGuard Asteroid Defense AI
# Requirements file

# Core data manipulation
pandas>=1.5.0
numpy>=1.23.0

# Machine learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
joblib>=1.2.0
pyyaml>=6.0

# Optional (for better performance)
scipy>=1.9.0

# For notebook
jupyter>=1.0.0
ipykernel>=6.19.0